BookPoq patch 0 2 (mongo)
