--- docker/docker-compose.yml
@@ -1,16 +1,0 @@
-version: '2'
-services:
-  rabbitmq:
-    image: rabbitmq:management
-    ports:
-    - 5672:5672
-    - 15672:15672
-    volumes:
-      - ./rabbitmq/definitions.json:/opt/rabbitmq/definitions.json:ro
-      - ./rabbitmq/rabbitmq.config:/etc/rabbitmq/rabbitmq.config:ro
-
-
-#
-#volumes:
-#  stubbornjava_esdata:
-#    driver: local
