--- docker/kafka/docker-compose.yml
@@ -1,19 +1,0 @@
-version: "2.1"
-
-services:
-  zookeeper:
-    image: artifactorycn.netcracker.com:17014/product/prod.platform.streaming_docker-zookeeper:v.3.5.5-0.2_20190624-073520
-    ports:
-      - 2181:2181
-
-  kafka:
-    image: artifactorycn.netcracker.com:17014/product/prod.platform.streaming_docker-kafka:v.2.2.1-0.1_20190625-101700
-    ports:
-      - 9092:9092
-      - 9093:9093
-    environment:
-      ZOOKEEPER_CONNECT: zookeeper:2181
-      HOST_NAME: "127.0.0.1"
-    depends_on:
-      zookeeper:
-        condition: service_started
