+++ pom.xml
@@ -10,10 +10,11 @@
     </parent>
 
 
-    <groupId>com.mykyta</groupId>
-    <artifactId>testspring</artifactId>
+    <groupId>com.book.poq</groupId>
+    <artifactId>book-poq</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>book-poq</name>
+    <packaging>jar</packaging>
 
     <properties>
         <java.version>1.8</java.version>
@@ -28,11 +29,6 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-jdbc</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-amqp</artifactId>
         </dependency>
 
         <dependency>
@@ -51,5 +47,14 @@
         </dependency>
 
     </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-maven-plugin</artifactId>
+            </plugin>
+        </plugins>
+    </build>
 
 </project>
