+++ src/main/java/com/book/poq/BookPoqApplication.java
@@ -1,0 +1,13 @@
+package com.book.poq;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BookPoqApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BookPoqApplication.class, args);
+    }
+
+}
