+++ src/main/java/com/book/poq/adapters/AdvertAdapter.java
@@ -1,0 +1,8 @@
+package com.book.poq.adapters;
+
+import com.book.poq.model.Advert;
+import reactor.core.publisher.Mono;
+
+public interface AdvertAdapter {
+    Mono<Advert> getAdvert(int id);
+}
