+++ src/main/java/com/book/poq/configurations/advert/AdvertProperties.java
@@ -1,0 +1,16 @@
+package com.book.poq.configurations.advert;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties("poq.advert")
+public class AdvertProperties {
+    private String url;
+
+    public String getUrl() {
+        return url;
+    }
+
+    public void setUrl(String url) {
+        this.url = url;
+    }
+}
