+++ src/main/java/com/book/poq/listeners/BookMessageListener.java
@@ -1,0 +1,7 @@
+package com.book.poq.listeners;
+
+import com.book.poq.model.Book;
+
+public interface BookMessageListener {
+    void process(Book bookMessage);
+}
