+++ src/main/java/com/book/poq/model/Advert.java
@@ -1,0 +1,38 @@
+package com.book.poq.model;
+
+public class Advert {
+    private int id;
+    private String text;
+
+    public Advert() {
+    }
+
+    public Advert(int id, String text) {
+        this.id = id;
+        this.text = text;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getText() {
+        return text;
+    }
+
+    public void setText(String text) {
+        this.text = text;
+    }
+
+    @Override
+    public String toString() {
+        return "Advert{" +
+                "id=" + id +
+                ", text='" + text + '\'' +
+                '}';
+    }
+}
