+++ src/main/java/com/book/poq/model/Book.java
@@ -1,0 +1,65 @@
+package com.book.poq.model;
+
+import java.io.Serializable;
+
+public class Book implements Serializable {
+    private int id;
+    private String title;
+    private String author;
+    private Advert advert;
+
+    public Book() {
+    }
+
+    public Book(
+            int id,
+            String title,
+            String author
+    ) {
+        this.id = id;
+        this.title = title;
+        this.author = author;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getAuthor() {
+        return author;
+    }
+
+    public void setAuthor(String author) {
+        this.author = author;
+    }
+
+    public Advert getAdvert() {
+        return advert;
+    }
+
+    public void setAdvert(Advert advert) {
+        this.advert = advert;
+    }
+
+    @Override
+    public String toString() {
+        return "Book{" +
+                "id=" + id +
+                ", title='" + title + '\'' +
+                ", author='" + author + '\'' +
+                ", advert=" + ((advert != null)? advert.getText() : "null") +
+                '}';
+    }
+}
