+++ src/main/java/com/book/poq/services/BookValidatorService.java
@@ -1,0 +1,8 @@
+package com.book.poq.services;
+
+import com.book.poq.model.Book;
+
+public interface BookValidatorService {
+
+    boolean validate(Book book);
+}
