--- src/main/java/com/mykyta/testspring/TestspringApplication.java
@@ -1,13 +1,0 @@
-package com.mykyta.testspring;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TestspringApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TestspringApplication.class, args);
-    }
-
-}
