--- src/main/java/com/mykyta/testspring/listeners/BookMessageListener.java
@@ -1,7 +1,0 @@
-package com.mykyta.testspring.listeners;
-
-import com.mykyta.testspring.model.Book;
-
-public interface BookMessageListener {
-    void process(Book bookMessage);
-}
