--- src/main/java/com/mykyta/testspring/services/BookValidatorService.java
@@ -1,8 +1,0 @@
-package com.mykyta.testspring.services;
-
-import com.mykyta.testspring.model.Book;
-
-public interface BookValidatorService {
-
-    boolean validate(Book book);
-}
